only make openbox config backup if there is no backup

+++ modules/openbox/install.sh
@@ -4,8 +4,10 @@ DOTFILES=~/dotfiles
 
 mkdir -p ~/.config/openbox
 for f in $DOTFILES/modules/openbox/openbox/*; do
-    if [ -f ~/.config/openbox/$(basename $f) ]; then
-        cp ~/.config/openbox/$(basename $f) ~/.config/openbox/$(basename $f)-original
+    if [ ! -f ~/.config/openbox/$(basename $f)-original ]; then
+        if [ -f ~/.config/openbox/$(basename $f) ]; then
+            cp ~/.config/openbox/$(basename $f) ~/.config/openbox/$(basename $f)-original
+        fi
     fi
     cp $f ~/.config/openbox/$(basename $f)
 done
